refactor: step1 리뷰 반영
- 구현부 /domain, /annotation 패키지 하위에 구현
- @DisplayName 을 통해 테스트 목적 명시
- Assertions.assertAll를 통해 모든 검증문 실행하도록 수정
