+++ src/main/java/annotation/PrintView.java
@@ -1,4 +1,4 @@
-package persistence.study;
+package annotation;
 
 public @interface PrintView {
 }
