+++ src/main/java/persistence/sql/ddl/QueryBuilder.java
@@ -1,0 +1,4 @@
+package persistence.sql.ddl;
+
+public class QueryBuilder {
+}
