+++ src/test/java/persistence/study/ReflectionTest.java
@@ -1,16 +1,19 @@
 package persistence.study;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import domain.Car;
+import annotation.PrintView;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
 import java.util.Arrays;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 class ReflectionTest {
     private static final Logger logger = LoggerFactory.getLogger(ReflectionTest.class);
 
@@ -22,7 +25,8 @@ class ReflectionTest {
     }
 
     @Test
-    void testMethodRun() throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
+    @DisplayName("Car 객체의 test로 시작하는 메소드를 실행하기")
+    void testMethodRun() throws Exception {
         Class<Car> carClass = Car.class;
         Car car = carClass.getConstructor().newInstance();
         Arrays.stream(carClass.getDeclaredMethods())
@@ -37,7 +41,8 @@ class ReflectionTest {
     }
 
     @Test
-    void testAnnotationMethodRun() throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
+    @DisplayName("Car 객체의 @PrintView 설정되어 있는 메소드를 실행하기")
+    void testAnnotationMethodRun() throws Exception {
         Class<Car> carClass = Car.class;
         Car car = carClass.getConstructor().newInstance();
         Arrays.stream(carClass.getDeclaredMethods())
@@ -52,7 +57,8 @@ class ReflectionTest {
     }
 
     @Test
-    public void privateFieldAccess() throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException, NoSuchFieldException {
+    @DisplayName("Car 객체의 private 필드에 값을 할당 후 값을 확인하기")
+    public void privateFieldAccess() throws Exception {
         Class<Car> clazz = Car.class;
         logger.debug(clazz.getName());
 
@@ -63,16 +69,20 @@ class ReflectionTest {
         Field nameField = clazz.getDeclaredField("name");
         nameField.setAccessible(true);
         nameField.set(car, name);
-        assertThat(car.getName()).isEqualTo(name);
 
         Field priceField = clazz.getDeclaredField("price");
         priceField.setAccessible(true);
         priceField.set(car, price);
-        assertThat(car.getPrice()).isEqualTo(price);
+
+        Assertions.assertAll(
+                () -> assertTrue(car.getName().equals(name))
+                , () -> assertTrue(car.getPrice() == price)
+        );
     }
 
     @Test
-    void constructorWithArgs() throws InvocationTargetException, InstantiationException, IllegalAccessException {
+    @DisplayName("Reflection API를 활용해 Car 인스턴스를 생성하기")
+    void constructorWithArgs() throws Exception {
         Class<Car> clazz = Car.class;
         String name = "test_name";
         int price = 0;
@@ -82,8 +92,10 @@ class ReflectionTest {
                 .findFirst().orElseThrow();
         Car car = (Car) constructor.newInstance(name, price);
 
-        assertThat(car.getName()).isEqualTo(name);
-        assertThat(car.getPrice()).isEqualTo(price);
+        Assertions.assertAll(
+                () -> assertTrue(car.getName().equals(name))
+                , () -> assertTrue(car.getPrice() == price)
+        );
     }
 
 }
